refactor: AppContext에서 ThemeContext 분리
