+++ packages/assignment/src/@lib/context/index.ts
@@ -1,1 +1,4 @@
+export * from './AppContext';
+export * from './ThemeContext';
 export * from './useAppContext';
+export * from './useThemeContext';
