+++ packages/assignment/src/@lib/context/useThemeContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ThemeContext } from './ThemeContext';
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useThemeContext must be used within a ThemeProvider');
+  }
+  return context;
+};
